refactor: move UserReviews into its own dedicated file
